back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Meeting Alignment AI FastAPI backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
